move: 전역적인 설정 클래스 패키지 세분화
