+++ src/main/java/com/koddy/server/global/config/infra/RestTemplateConfiguration.java
@@ -1,4 +1,4 @@
-package com.koddy.server.global.config;
+package com.koddy.server.global.config.infra;
 
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
