+++ src/test/java/com/koddy/server/common/config/TestWebBeanConfiguration.java
@@ -1,7 +1,7 @@
 package com.koddy.server.common.config;
 
 import com.koddy.server.auth.utils.TokenResponseWriter;
-import com.koddy.server.global.config.properties.CorsProperties;
+import com.koddy.server.global.config.web.CorsProperties;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
